small correction
In "number-game-errors.html" file and section " ###SyntaxError: missing ; before statement" the userGuess variable is defined as const and at line 195 and 201 userGuess variable is defined as let. In this topic, the error discussed related to equality operator. The userGuess must be const according to file "number-game-errors.html" but it is defined as let.